CSS highlight pseudos: parse ::search-text and ::search-text:current
This patch adds a new runtime enabled feature for ::search-text
(SearchTextHighlightPseudo), new PseudoId and PseudoType and devtools
protocol variants for ::search-text and :current, and extends the
selector parser to accept ::search-text and ::search-text:current
when the feature is enabled.

:current has an alternate functional form [1] that, at least in a
WebVTT context, allows you to match ancestors of the current element.
Parsing this will be left to a separate patch.

[1] https://drafts.csswg.org/selectors/#the-current-pseudo

Bug: 339298411
Change-Id: I2f12acd9580ae03dbdd816a03cbb68017864d136